Move java classes registered in Scala 3 compiler to scala-2 sources. Create renamed versions for Scala 3
