+++ javalib/src/main/scala-2/java/lang/ThrowablesCompat.scala
@@ -1,0 +1,8 @@
+// Classes in this file needs special handling in Scala 3, we need to make sure
+// that they would not be compiled with Scala 3 compiler
+
+package java.lang
+
+class NullPointerException(s: String) extends RuntimeException(s) {
+  def this() = this(null)
+}
