+++ javalib/src/main/scala-3/java/lang/ThrowablesCompat.scala
@@ -1,0 +1,9 @@
+// Classes defined in this file are registered inside Scala 3 compiler,
+// compiling them in javalib would lead to fatal error of compiler. They need
+// to be defined with a different name and renamed when generating NIR name
+
+package java.lang
+
+class _NullPointerException(s: String) extends RuntimeException(s) {
+  def this() = this(null)
+}
