+++ javalib/src/main/scala-3/java/lang/annotation/Retention.scala
@@ -1,3 +1,3 @@
 package java.lang.annotation
 
-trait Retention
+trait _Retention
