--- javalib/src/main/scala/java/io/Serializable.scala
@@ -1,3 +1,0 @@
-package java.io
-
-trait Serializable
