--- javalib/src/main/scala/java/lang/Cloneable.scala
@@ -1,3 +1,0 @@
-package java.lang
-
-trait Cloneable
