--- javalib/src/main/scala/java/lang/Comparable.scala
@@ -1,5 +1,0 @@
-package java.lang
-
-trait Comparable[A] {
-  def compareTo(o: A): scala.Int
-}
