+++ javalib/src/main/scala/java/lang/Throwables.scala
@@ -473,10 +473,6 @@ class NoSuchMethodException(s: String) extends ReflectiveOperationException(s) {
   def this() = this(null)
 }
 
-class NullPointerException(s: String) extends RuntimeException(s) {
-  def this() = this(null)
-}
-
 class NumberFormatException(s: String) extends IllegalArgumentException(s) {
   def this() = this(null)
 }
